本番環境でcustom.jsが読み込まれないため、custom.jsの削除及びその内容をapplication.jsへ移行
